Upgrade to the latest version of the Firebase SDK.
We're still relying on the "compatibility" API.

This also changes how we integrate with firecrypt.  Instead of
relying on monkeypatching we offer a new hook for wrapping the
Database object before it's used for anything.  The hook can only
be set before any messages are processed, so if you set it
successfully you are guaranteed that all database access will
go through the wrapper.