Guard against complex error properties and `postMessage` failures.
This attempts to address the issue reported in
https://reviewable.slack.com/archives/C07T91SN8FL/p1764006291030979?thread_ts=1763429032.020499&cid=C07T91SN8FL.

1. Only copy over primitive properties when converting an error
to JSON before sending it up to the client.  I suspect that the
Firebase `App` object is somehow getting included in an error,
causing the data cloning issue referenced above.  In theory,
Firebase errors are only supposed to have top-level primitive
properties, so we shouldn't lose any information by restricting
ourselves to those.

2. In case that wasn't the cause of the issue, also add extra
guards around `postMessage` so that errors will be caught and
retransmitted to the client, where they can cause a crash rather
than having a Firebase call get stuck.  We're limited in what
useful information we can include at this layer but at least
there'll be a clear signal that something went terribly wrong.